% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-files.R
\name{install_codespace_files}
\alias{install_codespace_files}
\title{Install codespace files into a '.devcontainer' directory at this location}
\usage{
install_codespace_files(path = ".")
}
\arguments{
\item{path}{path to install into}
}
\value{
TRUE
}
\description{
Install codespace files into a '.devcontainer' directory at this location
}
